add barra de carregamento no login e cadastro, junto a correção do redirect
